fix(web-scraper): table of contents
We are now generating custom table of contents from the readme content and we might want to remove the one that is within `<!-- toc -->` and `<!-- tocstop -->` (we do that on web as well). I just moved the Table of contents heading inside the comment tags so that it is removed as well.